fix(types): relocate D1 layout declarations (SD-2893)
