+++ packages/super-editor/src/editors/v1/core/presentation-editor/layout/FootnotesBuilder.ts
@@ -20,7 +20,8 @@
 
 import type { EditorState } from 'prosemirror-state';
 import type { FlowBlock } from '@superdoc/contracts';
-import { toFlowBlocks, type ConverterContext } from '@superdoc/pm-adapter';
+import { toFlowBlocks } from '@superdoc/pm-adapter';
+import type { ConverterContext } from '@superdoc/pm-adapter/converter-context.js';
 import { SUBSCRIPT_SUPERSCRIPT_SCALE } from '@superdoc/pm-adapter/constants.js';
 
 import type { ProseMirrorJSON } from '../../types/EditorTypes.js';
